workspace/requirements.txt
pyserial
pyusb
pycryptodome
lxml
colorama
